<fix> Configure Global exception function
This commit adds EntityNotFoundException, DuplicateKeyException in Global exception.
