+++ src/main/java/vn/unigap/api/controller/EmployerController.java
@@ -32,97 +32,48 @@ public class EmployerController {
 
     // create employer
     @PostMapping
-    public ResponseEntity<?> createEmployer(@Valid @RequestBody CreateEmployerDtoIn employer) {
-        try {
-            String responseService = employerService.createEmployer(employer);
+    public ResponseEntity<?> createEmployer(@Valid @RequestBody CreateEmployerDtoIn employer) throws DuplicateKeyException, EntityNotFoundException {
+        String responseService = employerService.createEmployer(employer);
 
-            return ResponseEntity.status(HttpStatus.CREATED)
-                    .body(new CustomResponse<>(0, HttpStatus.CREATED.value(), responseService, ""));
-        } catch (EntityNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(new CustomResponse<>(1, HttpStatus.NOT_FOUND.value(), e.getMessage(), ""));
-        } catch (DuplicateKeyException e) {
-            log.error("An error occurred at controller", e);
-            return ResponseEntity.status(HttpStatus.CONFLICT)
-                    .body(new CustomResponse<>(2, HttpStatus.CONFLICT.value(), e.getMessage(), ""));
-        } catch (Exception e) {
-            log.error("An error occurred at service", e);
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(new CustomResponse<>(5, HttpStatus.INTERNAL_SERVER_ERROR.value(), e.getMessage(), ""));
-        }
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body(new CustomResponse<>(0, HttpStatus.CREATED.value(), responseService, ""));
     }
 
     // get all employers
     @GetMapping
     public ResponseEntity<?> getAllEmployer(@Valid @ModelAttribute PaginateEmployerDtoIn paginateEmployerDtoIn) {
-        try {
-            PaginationResponse responseService = employerService.getAllEmployers(paginateEmployerDtoIn);
+        PaginationResponse responseService = employerService.getAllEmployers(paginateEmployerDtoIn);
 
-            return ResponseEntity.status(HttpStatus.OK)
-                    .body(new CustomResponse<>(0, HttpStatus.OK.value(), "Get " +
-                            "employer list successfully", responseService));
-        } catch (Exception e) {
-            log.info("An error occurred at controller", e);
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(new CustomResponse<>(5, HttpStatus.INTERNAL_SERVER_ERROR.value(), e.getMessage(), ""));
-        }
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(new CustomResponse<>(0, HttpStatus.OK.value(), "Get " +
+                        "employer list successfully", responseService));
     }
 
     // get employer by Id
     @GetMapping("/{id}")
-    public ResponseEntity<?> getUserById(@PathVariable Long id) {
-        try {
-            Object responseService = employerService.getEmployerById(id);
+    public ResponseEntity<?> getUserById(@PathVariable Long id) throws EntityNotFoundException {
+        Object responseService = employerService.getEmployerById(id);
 
-            return ResponseEntity.status(HttpStatus.OK).body(new CustomResponse<>(0, HttpStatus.OK.value(), "Get " +
-                    "employer by Id successful", responseService));
-        } catch (EntityNotFoundException e) {
-            log.error("An error occurred at controller", e);
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(new CustomResponse<>(1, HttpStatus.NOT_FOUND.value(), e.getMessage(), ""));
-        } catch (Exception e) {
-            log.error("An error occurred at controller", e);
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(new CustomResponse<>(5, HttpStatus.INTERNAL_SERVER_ERROR.value(), e.getMessage(), ""));
-        }
+        return ResponseEntity.status(HttpStatus.OK).body(new CustomResponse<>(0, HttpStatus.OK.value(), "Get " +
+                "employer by Id successful", responseService));
     }
 
     // update employer
     @PutMapping("/{id}")
     public ResponseEntity<?> updateEmployer(@PathVariable Long id,
-                                            @Valid @RequestBody UpdateEmployerDtoIn employer) {
-        try {
-            String responseService = employerService.updateEmployer(id, employer);
+                                            @Valid @RequestBody UpdateEmployerDtoIn employer) throws EntityNotFoundException {
+        String responseService = employerService.updateEmployer(id, employer);
 
-            return ResponseEntity.status(HttpStatus.OK)
-                    .body(new CustomResponse<>(0, HttpStatus.OK.value(), responseService, ""));
-        } catch (EntityNotFoundException e) {
-            log.error("check not found " + e);
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(new CustomResponse<>(1, HttpStatus.NOT_FOUND.value(), e.getMessage(), ""));
-        } catch (Exception e) {
-            log.error("An error occurred at controller", e);
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(new CustomResponse<>(5, HttpStatus.INTERNAL_SERVER_ERROR.value(), e.getMessage(), ""));
-        }
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(new CustomResponse<>(0, HttpStatus.OK.value(), responseService, ""));
     }
 
     // delete employer
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteEmployer(@PathVariable Long id) {
-        try {
-            String responseService = employerService.deleteEmployer(id);
+    public ResponseEntity<?> deleteEmployer(@PathVariable Long id) throws EntityNotFoundException {
+        String responseService = employerService.deleteEmployer(id);
 
-            return ResponseEntity.status(HttpStatus.OK)
-                    .body(new CustomResponse<>(0, HttpStatus.OK.value(), responseService, ""));
-        } catch (EntityNotFoundException e) {
-            log.error("check not found " + e);
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(new CustomResponse<>(1, HttpStatus.NOT_FOUND.value(), e.getMessage(), ""));
-        } catch (Exception e) {
-            log.error("An error occurred at controller", e);
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(new CustomResponse<>(5, HttpStatus.INTERNAL_SERVER_ERROR.value(), e.getMessage(), ""));
-        }
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(new CustomResponse<>(0, HttpStatus.OK.value(), responseService, ""));
     }
 }
