--- src/main/java/vn/unigap/api/controller/advice/GlobalExceptionHandler.java
@@ -1,34 +1,0 @@
-package vn.unigap.api.controller.advice;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import vn.unigap.api.dto.out.CustomResponse;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<?> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        FieldError errors = ex.getBindingResult().getFieldError();
-        assert errors != null;
-        String errorMessage = errors.getDefaultMessage();
-
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new CustomResponse<>(3,
-                HttpStatus.BAD_REQUEST.value(), errorMessage, ""));
-    }
-
-//    @ExceptionHandler(MethodArgumentNotValidException.class)
-//    public ResponseEntity<Map<String, String>> handleValidationExceptions(MethodArgumentNotValidException ex) {
-//        Map<String, String> errors = new HashMap<>();
-//        ex.getBindingResult().getFieldErrors().forEach((error) -> {
-//            String fieldName = error.getField();
-//            String errorMessage = error.getDefaultMessage();
-//            errors.put(fieldName, errorMessage);
-//        });
-//        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body( errors);
-//    }
-}
